`Trimesh.fill_holes` was adding new faces but haphazardly managing the
bookkeeping: it broke `len(mesh.faces) == len(mesh.face_attributes[n])`
if there were any holes filled.

- add `Trimesh.extend_faces` which moves the bookkeeping of "preserve
cached face normals to avoid recalculating, pad face colors, pad face
attributes" to a generic mesh attribute in the same vein as
`mesh.update_faces`.
- refactor the tesselation of the holes in `repair.fill_holes` to use
`triangulate_quads` which implemented the hole filling logic but better.
Switch the winding search to a numpy indexing trick for new faces.
- Have `geometry.triangulate_quads` use the preexisting
`util.triangle_fan_to_faces` which uses numpy instead of a list
comprehension and is heavily used and tested in the codebase.

Also:
- try to test and fix #2516

+++ tests/test_encoding.py
@@ -162,7 +162,7 @@ class EncodingTest(g.unittest.TestCase):
         box = g.trimesh.primitives.Box()
         box.apply_translation([0.5, 0.5, 0.5])  # center at origin
         box.apply_scale(5)  # 0 -> 5
-        expected_sparse_indices = np.array(box.vertices)
+        expected_sparse_indices = np.array(box.vertices, dtype=int)
         box.apply_translation([2, 2, 2])  # 2 -> 7
         sparse = np.array(box.vertices, dtype=int)
         encoding = enc.SparseBinaryEncoding(sparse, shape=(9, 9, 9))
